+++ react/javascript/src/components/gherkin/Scenario.tsx
@@ -21,7 +21,7 @@ const Scenario: React.FunctionComponent<IProps> = ({ scenario }) => {
   const cucumberQuery = React.useContext(CucumberQueryContext)
   const gherkinQuery = React.useContext(GherkinQueryContext)
   const uri = React.useContext(UriContext)
-  const pickleIds = gherkinQuery.getPickleIds(uri, scenario.location.line)
+  const pickleIds = gherkinQuery.getPickleIds(uri, scenario.id)
   const beforeHooks = cucumberQuery.getBeforeHookSteps(pickleIds[0])
   const afterHooks = cucumberQuery.getAfterHookSteps(pickleIds[0])